fix CRUD category, add soft delete for category model

+++ database/migrations/2022_01_19_074737_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->string('slug')->unique();
             $table->float('price')->default(0);
-            $table->integer('weight')->default(0);
+            $table->float('weight')->default(0);
             $table->integer('quantity')->default(false);
             $table->text('description')->nullable();
             $table->boolean('is_delete')->default(false);
